// node-projects/backups/nub/20250426/interface/types.ts
export type Flatten<T> = T extends Record<string, unknown>
  ? UnionToIntersection<
      {
        [K in keyof DeepRequired<T>]: K extends string
          ? T[K] extends Record<string, unknown>
            ? {
                [P in `${K}.${keyof Flatten<T[K]> & string}`]: Flatten<
                  T[K]
                >[P extends `${K}.${infer R}` ? R : never];
              } & { [P in K]: T[K] }
            : { [P in K]: T[K] }
          : never;
      }[keyof T]
    >
  : never;

export type UnionToIntersection<T> = (
  T extends unknown ? (x: T) => void : never
) extends (x: infer R) => void
  ? R
  : never;

/** Recursively make all properties of a map/object required */
export type DeepRequired<T> = {
  [K in keyof T]-?: Exclude<T[K], undefined> extends Record<string, unknown>
    ? DeepRequired<T[K]>
    : T[K];
};
// ...
export enum UtilityProduct {
  Airtime = "airtime",
  Prepaid = "prepaid",
  Postpaid = "postpaid",
  Plan = "plan",
  Default = "default",
}

// TODO: update
export type TPlanData = {
  providerId: ProviderId;
  planCode: string;
  amount: number;
};

export type TPaymentData = TUtilityMeta &
  (
    | {
        // Require plan data as well, if service is of 'plan' type...
        productId: UtilityProduct.Plan;
        planData: TPlanData;
      }
    | {
        // ...otherwise, it should not be present.
        productId: Exclude<UtilityProduct, "plan">;
        planData?: undefined;
      }
  );

export type TProcessData = {
  paymentData: TPaymentData;
  providerId: ProviderId;
  uid: string;
  appTxHash: string;
  nairaAdminDocPath: string;
  txPath: string;
};
// ...
export enum ProviderId {
  ProviderX = "providerX", // TODO: remove; test provider
}
// ...
export type TUtilityMeta = {
  /** Unique identifier for the customer, at the merchant */
  customerId: string;
  merchantId: UtilityMerchant;
  productId: UtilityProduct;
  planId?: string;
  amount?: number;
} & (
  | {
      // Require planId if service is of 'plan' type...
      productId: UtilityProduct.Plan;
      planId: string;
      amount?: undefined;
    }
  | {
      // ...else, require amount instead
      productId: Exclude<UtilityProduct, "plan">;
      planId?: undefined;
      amount: number;
    }
);